+++ ja2lib/Tactical/MercHiring.c
@@ -66,8 +66,8 @@ extern BOOLEAN gfFirstHeliRun;
 // ATE: Globals that dictate where the mercs will land once being hired
 // Default to Omerta
 // Saved in general saved game structure
-INT16 gsMercArriveSectorX = 9;
-INT16 gsMercArriveSectorY = 1;
+u8 gsMercArriveSectorX = 9;
+u8 gsMercArriveSectorY = 1;
 
 static void CheckForValidArrivalSector();
 
@@ -301,7 +301,7 @@ void MercArrivesCallback(UINT8 ubSoldierID) {
       // OK, If we are in mapscreen, get out...
       if (IsMapScreen_2()) {
         // ATE: Make sure the current one is selected!
-        ChangeSelectedMapSector(gWorldSectorX, gWorldSectorY, 0);
+        ChangeSelectedMapSector((u8)gWorldSectorX, (u8)gWorldSectorY, 0);
 
         RequestTriggerExitFromMapscreen(MAP_EXIT_TO_TACTICAL);
       }
